misc changes

+++ Simduino/NewPing.h
@@ -126,7 +126,7 @@
 #define PING_MEDIAN_DELAY 29    // Millisecond delay between pings in the ping_median method.
 
 // Conversion from uS to distance (round result to nearest cm or inch).
-#define NewPingConvert(echoTime, conversionFactor) (max((echoTime + conversionFactor / 2) / conversionFactor, (echoTime ? 1 : 0)))
+#define NewPingConvert(echoTime, conversionFactor) (max((echoTime + conversionFactor / 2) / conversionFactor, (unsigned int) (echoTime ? 1 : 0)))
 
 
 class NewPing {
